Backend/app/requirements.txt
fastapi
boto3
psycopg2-binary
mysql-connector-python
python-dotenv
uvicorn
pydantic